backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
